Created Catalog section
Created catalog page, modals, implemented search, sort, add, replace functions

+++ src/app/catalog/catalog.component.ts
@@ -1,15 +1,19 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import {
-  CatalogElement,
-  CustomerElement,
-  SectionDataInterface,
-} from '../shared/interfaces/intrefaces';
+import { CatalogElement } from '../shared/interfaces/intrefaces';
 import { MatTableDataSource } from '@angular/material/table';
 import { Store } from '@ngrx/store';
-import { changeSideBar } from '../shared/actions/side-bar.action';
+import { changeSideBar } from '../shared/store/actions/side-bar.action';
 import { AuthService } from '../shared/services/auth.service';
 import { MatPaginator } from '@angular/material/paginator';
 import { first } from 'rxjs/operators';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { EditCatalogModalComponent } from '../components/modals/edit-catalog-modal/edit-catalog-modal.component';
+import { DeleteCatalogModalComponent } from '../components/modals/delete-catalog-modal/delete-catalog-modal.component';
+import { FormControl, FormGroup } from '@angular/forms';
+import { AddCatalogModalComponent } from '../components/modals/add-catalog-modal/add-catalog-modal.component';
+import { CatalogService } from '../shared/services/catalog.service';
+import { ReplaceCatalogModalComponent } from '../components/modals/replace-catalog-modal/replace-catalog-modal.component';
+
 @Component({
   selector: 'app-catalog',
   templateUrl: './catalog.component.html',
@@ -17,7 +21,14 @@ import { first } from 'rxjs/operators';
 })
 export class CatalogComponent implements OnInit {
   @ViewChild('paginator') paginator: MatPaginator;
+  isDeleteDialogRef: MatDialogRef<DeleteCatalogModalComponent>;
   dataSource: MatTableDataSource<CatalogElement>;
+  availabilitySelectActive: boolean = false;
+  availabilitySelectForm = new FormGroup({
+    'In Stock': new FormControl(false),
+    'Out Stock': new FormControl(false),
+    Discontinued: new FormControl(false),
+  });
   displayedColumns: string[] = [
     'productCode',
     'productName',
@@ -26,18 +37,16 @@ export class CatalogComponent implements OnInit {
     'productAvailability',
     'action',
   ];
+  columnWidth: string = '16.666%';
 
-  constructor(private store: Store, private authService: AuthService) {}
+  constructor(
+    private store: Store,
+    private catalogService: CatalogService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-    this.authService
-      .getCatalog({ token: localStorage.getItem('token') })
-      .pipe(first())
-      .subscribe((res) => {
-        res = this.concatProductAndUnits(res);
-        this.dataSource = new MatTableDataSource<any>(res.products);
-        this.dataSource.paginator = this.paginator;
-      });
+    this.getSortCatalog();
   }
 
   concatProductAndUnits(res: any): any {
@@ -50,13 +59,31 @@ export class CatalogComponent implements OnInit {
     return res;
   }
 
-  openEditModal(el: any): void {}
+  openEditModal(el: any): void {
+    const a = this.dialog.open(EditCatalogModalComponent, {
+      data: el,
+    });
+    a.afterClosed()
+      .pipe(first())
+      .subscribe((res) => {
+        setTimeout(() => this.getSortCatalog(), 1000);
+      });
+  }
 
-  getSortCatalog(): void {}
+  getSortCatalog(): void {
+    this.catalogService
+      .getCatalog({ token: localStorage.getItem('token') })
+      .pipe(first())
+      .subscribe((res) => {
+        res = this.concatProductAndUnits(res);
+        this.dataSource = new MatTableDataSource<any>(res.products);
+        this.dataSource.paginator = this.paginator;
+      });
+  }
 
   searchCatalog(event: any): void {
     if (event.target.value === '') {
-      this.authService
+      this.catalogService
         .getCatalog({ token: localStorage.getItem('token') })
         .subscribe((res) => {
           res = this.concatProductAndUnits(res);
@@ -64,23 +91,87 @@ export class CatalogComponent implements OnInit {
         });
       return;
     }
-    this.authService.searchCatalog(event.target.value).subscribe((res) => {
+    this.catalogService.searchCatalog(event.target.value).subscribe((res) => {
       res = this.concatProductAndUnits(res);
       this.dataSource = new MatTableDataSource<any>(res.products);
       this.dataSource.paginator = this.paginator;
     });
   }
 
-  deleteCatalogEl(id: number): void {
-    this.authService
-      .deleteCatalogEl(id)
+  deleteCatalogEl(id: number, name: string): void {
+    this.isDeleteDialogRef = this.dialog.open(DeleteCatalogModalComponent, {
+      data: name,
+    });
+    this.isDeleteDialogRef
+      .afterClosed()
       .pipe(first())
       .subscribe((res) => {
-        this.getSortCatalog();
+        if (res != undefined && res.isDelete) {
+          if (res.isDelete) {
+            this.catalogService
+              .deleteCatalogEl(id)
+              .pipe(first())
+              .toPromise()
+              .then((data) => {
+                this.getSortCatalog();
+              });
+          }
+        }
       });
   }
 
   changeSideBarState(): void {
     this.store.dispatch(changeSideBar());
+  }
+
+  toggleAvailabilitySelect(): void {
+    this.availabilitySelectActive = !this.availabilitySelectActive;
+  }
+
+  sortCatalog(): void {
+    let sortParamsArray = Object.entries(this.availabilitySelectForm.value)
+      .filter((el) => {
+        return el[1];
+      })
+      .map((el) => {
+        return el[0];
+      });
+    if (sortParamsArray.length === 0) {
+      this.catalogService
+        .getCatalog({ token: localStorage.getItem('token') })
+        .subscribe((res) => {
+          res = this.concatProductAndUnits(res);
+          this.dataSource = new MatTableDataSource<any>(res.products);
+        });
+      return;
+    }
+    this.catalogService.sortCatalog(sortParamsArray).subscribe((res) => {
+      res = this.concatProductAndUnits(res);
+      this.dataSource = new MatTableDataSource<any>(res.products);
+      this.dataSource.paginator = this.paginator;
+    });
+  }
+
+  openAddModal(): void {
+    const a = this.dialog.open(AddCatalogModalComponent);
+    a.afterClosed()
+      .pipe(first())
+      .subscribe((res) => {
+        setTimeout(() => this.getSortCatalog(), 1000);
+      });
+  }
+
+  replaceCatalog(): void {
+    const a = this.dialog.open(ReplaceCatalogModalComponent);
+    a.afterClosed()
+      .pipe(first())
+      .toPromise()
+      .then((data) => {
+        if (data != undefined && data.isReplace) {
+          if (data.isReplace) {
+            setTimeout(() => this.getSortCatalog(), 1000);
+          }
+        }
+      });
   }
 }
